+++ build-system/app-index/header-links.js
@@ -1,0 +1,44 @@
+/**
+ * Copyright 2018 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+module.exports = [
+  {
+    'name': 'Developing',
+    'href': 'https://' +
+      'github.com/ampproject/amphtml/blob/master/contributing/DEVELOPING.md',
+  },
+  {
+    'divider': true,
+    'name': 'Contributing',
+    'href': 'https://github.com/ampproject/amphtml/blob/master/CONTRIBUTING.md',
+  },
+  {
+    'name': 'Github',
+    'href': 'https://github.com/ampproject/amphtml/',
+  },
+  {
+    'name': 'Find File',
+    'href': 'https://github.com/ampproject/amphtml/find/master',
+  },
+  {
+    'name': 'Travis',
+    'href': 'https://travis-ci.org/ampproject/amphtml',
+  },
+  {
+    'name': 'Percy',
+    'href': 'https://percy.io/ampproject/amphtml/',
+  },
+];
